+++ pcsx2/GS/Renderers/HW/GSTextureCache.h
@@ -242,7 +242,7 @@ public:
 		bool m_valid_rgb = false;
 		bool m_rt_alpha_scale = false;
 		bool m_downscaled = false;
-		int m_last_draw = 0;
+		u64 m_last_draw = 0;
 
 		bool m_is_frame = false;
 		bool m_used = false;
@@ -452,7 +452,7 @@ protected:
 	TempZAddress m_temporary_z_info;
 
 	Target* m_last_clut_target = nullptr;
-	int m_last_gpu_clut_draw = 0;
+	u64 m_last_gpu_clut_draw = 0;
 
 	std::unique_ptr<GSDownloadTexture> m_color_download_texture;
 	std::unique_ptr<GSDownloadTexture> m_uint16_download_texture;
@@ -494,7 +494,7 @@ public:
 	__fi u64 GetTotalHashCacheMemoryUsage() const { return (m_hash_cache_memory_usage + m_hash_cache_replacement_memory_usage); }
 	__fi u64 GetSourceMemoryUsage() const { return m_source_memory_usage; }
 	__fi u64 GetTargetMemoryUsage() const { return m_target_memory_usage; }
-	__fi int GetLastGPUCLUTDraw() { return m_last_gpu_clut_draw; };
+	__fi u64 GetLastGPUCLUTDraw() { return m_last_gpu_clut_draw; };
 
 	void Read(Target* t, const GSVector4i& r);
 	void Read(Source* t, const GSVector4i& r);
